html/semantics/forms/the-input-element/month.html

<!DOCTYPE html>
<html>

  <head>
    <title>Inputs Month</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Morishita Hiromitsu" href="mailto:hero@asterisk-works.jp">
	  <link rel="help" href="http://www.w3.org/TR/html5/the-input-element.html#the-input-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>

  		<h1>Inputs Month</h1>
  		<div style="display: none">
	  		<input id="valid" type="month" value="2011-1" min="2011-01" max="2011-12" />
        <input id="too_small_value" type="month" value="1999-01" min="2011-01" max="2011-12"/>
        <input id="too_large_value" type="month" value="2099-01" min="2011-01" max="2011-12"/>
        <input id="invalid_value" type="month" value="invalid-month" min="2011-01" max="2011-12"/>
        <input id="invalid_min" type="month" value="2011-01" min="invalid_min" max="2011-12"/>
        <input id="invalid_max" type="month" value="2011-01" min="2011-01" max="invalid_max"/>
        <input id="min_larger_than_max" type="month" value="2011-01" min="2099-01" max="2011-12"/>
		</div>
		
    <div id="log">
    </div>
	
	<script type="text/javascript">

		test(function() {assert_equals(document.getElementById("valid").type, "month")},
        "month type support on input element", {
		  "help" : "http://www.w3.org/html/wg/drafts/html/master/forms.html#month-state-(type=month)" });

		test(function() {
          assert_equals(document.getElementById("valid").value, "2011-01"),
          assert_equals(document.getElementById("too_small_value").value, "2011-01"),
          assert_equals(document.getElementById("too_large_value").value, "2011-12")},
        "The value attribute, if specified and not empty, must have a value that is a valid month string", {
		  "help" : "http://www.w3.org/html/wg/drafts/html/master/forms.html#month-state-(type=month)" });

		test(function() {
          assert_equals(document.getElementById("valid").min, "2011-01"),
          assert_equals(document.getElementById("invalid_min").min, "")
          },
        "The min attribute, if specified, must have a value that is a valid month string.", {
		  "help" : "http://www.w3.org/html/wg/drafts/html/master/forms.html#month-state-(type=month)" });

		test(function() {
          assert_equals(document.getElementById("valid").max, "2011-12"),
          assert_equals(document.getElementById("min_larger_than_max").max, "2099-01"),
          assert_equals(document.getElementById("invalid_max").max, "")
        },
        "The max attribute, if specified, must have a value that is a valid month string", {
		  "help" : "http://www.w3.org/html/wg/drafts/html/master/forms.html#month-state-(type=month)" });

    test(function() {
          assert_equals(document.getElementById("invalid_value").value, "")
        },
        "User agents must not allow the user to set the value to a non-empty string that is not a valid month string.", {
          "help" : "http://www.w3.org/html/wg/drafts/html/master/forms.html#month-state-(type=month)" });
	</script>
	
  </body>
  
</html>
